--- a/test/clj/rems/browser_test_util.clj
+++ b/test/clj/rems/browser_test_util.clj
@@ -174,7 +174,7 @@
 ;; Etaoin's fill-human almost works, but very rarely loses characters,
 ;; probably due to the lack of a _minimum_ delay between keypresses.
 ;; This is a reimplementation.
-(def +character-delay+ 0.02)
+(def +character-delay+ 0.01)
 (def +max-extra-delay+ 0.2)
 (def +typo-probability+ 0.05)