baseline (master):
timings on my machine for `lein kaocha test-browser`: 476s, 496s, 492s
timings on CircleCI for `browser-test`: 10:27, 11:10, 9:58

after first commit:
my machine: 401s, 408s, 402s
CircleCI: 8:45, 8:05, 8:33

after second commit:
my machine: 371s, 395s, 435s
CircleCI: 8:04, 7:57, (timeout while starting driver), 8:11

after third commit:
my machine: 411s, 441s, 445s (weird, it should be faster, I guess my machine has more load now than previously)
CircleCI: 7:57, 7:46, (timeout while starting driver). Run the next day: 11:57, 11:10